testJMY

+++ src/tasks/tasks.controller.ts
@@ -2,33 +2,50 @@ import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/commo
 import { TasksService } from './tasks.service';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { UpdateTaskDto } from './dto/update-task.dto';
+import { Task } from '@prisma/client';
 
 @Controller('tasks')
 export class TasksController {
-  constructor(private readonly tasksService: TasksService) {}
+    constructor(private readonly tasksService: TasksService) { }
 
-  @Post()
-  create(@Body() createTaskDto: CreateTaskDto) {
-    return this.tasksService.create(createTaskDto);
-  }
+    @Post()
+    async create(@Body() createTaskDto: CreateTaskDto): Promise<Task> {
+        return await this.tasksService.create(createTaskDto);
+    }
 
-  @Get()
-  findAll() {
-    return this.tasksService.findAll();
-  }
+    @Get()
+    async findAll(): Promise<Task[]>  {
+        return await this.tasksService.findAll();
+    }
 
-  @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.tasksService.findOne(+id);
-  }
+    @Get(':id')
+   async findOne(@Param('id') id: string): Promise<Task | { message: string }> {
 
-  @Patch(':id')
-  update(@Param('id') id: string, @Body() updateTaskDto: UpdateTaskDto) {
-    return this.tasksService.update(+id, updateTaskDto);
-  }
+        const TaskId = await this.tasksService.findOne(+id)
+        if (!TaskId) {
+            return { message: "Cette tâche n'existe pas" }
+        }
+        return this.tasksService.findOne(+id);
+    }
 
-  @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.tasksService.remove(+id);
-  }
+    @Patch(':id')
+    async update(@Param('id') id: string, @Body() data: UpdateTaskDto): Promise<Task | { message: string }> {
+
+        const TaskId = await this.tasksService.findOne(+id)
+        if (!TaskId) {
+            return { message: "Cette tâche n'existe pas" }
+        }
+        return await this.tasksService.update(+id, data);
+    }
+
+    @Delete(':id')
+    async delete(@Param('id') id: string): Promise<{ message: string }> {
+        
+        const TaskId = await this.tasksService.findOne(+id)
+        if (!TaskId) {
+            return { message: "Cette tâche n'existe pas" }
+        }
+        await this.tasksService.remove(+id);
+        return { message: "Tâche supprimée" }
+    }
 }
