+++ src/user-task-events/user-task-events.controller.ts
@@ -2,33 +2,49 @@ import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/commo
 import { UserTaskEventsService } from './user-task-events.service';
 import { CreateUserTaskEventDto } from './dto/create-user-task-event.dto';
 import { UpdateUserTaskEventDto } from './dto/update-user-task-event.dto';
-
+import { UserTaskEvent } from '@prisma/client';
 @Controller('user-task-events')
 export class UserTaskEventsController {
-  constructor(private readonly userTaskEventsService: UserTaskEventsService) {}
+    constructor(private readonly userTaskEventsService: UserTaskEventsService) { }
 
-  @Post()
-  create(@Body() createUserTaskEventDto: CreateUserTaskEventDto) {
-    return this.userTaskEventsService.create(createUserTaskEventDto);
-  }
+    @Post()
+    async create(@Body() createUserTaskEventDto: CreateUserTaskEventDto): Promise<UserTaskEvent> {
+        return await this.userTaskEventsService.create(createUserTaskEventDto);
+    }
 
-  @Get()
-  findAll() {
-    return this.userTaskEventsService.findAll();
-  }
+    @Get()
+    async findAll(): Promise<UserTaskEvent[]> {
+        return await this.userTaskEventsService.findAll();
+    }
 
-  @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.userTaskEventsService.findOne(+id);
-  }
+    @Get(':id')
+    async findOne(@Param('id') id: string): Promise<UserTaskEvent | { message: string }> {
 
-  @Patch(':id')
-  update(@Param('id') id: string, @Body() updateUserTaskEventDto: UpdateUserTaskEventDto) {
-    return this.userTaskEventsService.update(+id, updateUserTaskEventDto);
-  }
+        const UserTaskEventId = await this.userTaskEventsService.findOne(+id)
+        if (!UserTaskEventId) {
+            return { message: " UserTaskEvent Not found" }
+        }
+        return await this.userTaskEventsService.findOne(+id);
+    }
 
-  @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.userTaskEventsService.remove(+id);
-  }
+    @Patch(':id')
+    async update(@Param('id') id: string, @Body() updateUserTaskEventDto: UpdateUserTaskEventDto): Promise<UserTaskEvent | { message: string }> {
+
+        const UserTaskEventId = await this.userTaskEventsService.findOne(+id)
+        if (!UserTaskEventId) {
+            return { message: " UserTaskEvent Not found" }
+        }
+        return this.userTaskEventsService.update(+id, updateUserTaskEventDto);
+    }
+
+    @Delete(':id')
+     async delete (@Param('id') id: string): Promise<UserTaskEvent | { message: string }>  {
+
+        const UserTaskEventId = await this.userTaskEventsService.findOne(+id)
+        if (!UserTaskEventId) {
+            return { message: " UserTaskEvent Not found" }
+        }
+        await this.userTaskEventsService.remove(+id);
+        return {message:"UserTaskEvent deleted"}
+    }
 }
